// client/one/app/core/users/users.constants.ts
export enum EntityPermissionValue {
    READ = 'read',
    WRITE = 'write',
    USER = 'user',
    BUDGET = 'budget',
    BUDGET_MARGIN = 'budget_margin',
    AGENCY_SPEND_MARGIN = 'agency_spend_margin',
    MEDIA_COST_DATA_COST_LICENCE_FEE = 'media_cost_data_cost_licence_fee',
    BASE_COSTS_SERVICE_FEE = 'base_costs_service_fee',
}